ui: switch to single file mode with snippet maximize
* switch back to snippet view with snippet button in file title bar
* use new title bar widget in single view as well
* restore view mode on undo, store current view mode for every message id (added id to message)
* treat view mode switch as adapt message on undo stack
* fixed crash after snippet expansion
* fixed hatching in file title not shown

+++ src/lib_gui/qt/element/QtCodeArea.h
@@ -79,7 +79,7 @@ public:
 	Id getLocationIdOfFirstActiveLocation(Id tokenId) const;
 	Id getLocationIdOfFirstActiveScopeLocation(Id tokenId) const;
 
-	uint getActiveLocationCount() const;
+	size_t getActiveLocationCount() const;
 
 	QRectF getLineRectForLineNumber(uint lineNumber) const;
